this is a mess, need to clean up
