+++ feagi_data_structures/src/genomic/cortical_area/cortical_type.rs
@@ -1,0 +1,15 @@
+use crate::genomic::cortical_area::io_cortical_area_data_type::IOCorticalAreaDataType;
+
+pub enum CorticalType {
+    Custom,
+    Memory,
+    Core(CoreCorticalType),
+    BrainInput(IOCorticalAreaDataType),
+    BrainOutput(IOCorticalAreaDataType)
+}
+
+
+pub enum CoreCorticalType {
+    Power,
+    Death
+}
