+++ feagi_data_structures/src/genomic/cortical_area/cortical_unit_type.rs
@@ -1,0 +1,9 @@
+use crate::genomic::cortical_area::io_cortical_area_data_type_descriptors::{FrameChangeHandling, PercentageNeuronPositioning};
+
+pub enum SensoryCorticalUnit {
+    Infrared(( (FrameChangeHandling, PercentageNeuronPositioning) )),
+    Vision(( (FrameChangeHandling, ) )),
+    SegmentedVision((FrameChangeHandling, ), (FrameChangeHandling, ), (FrameChangeHandling, ),
+                    (FrameChangeHandling, ), (FrameChangeHandling, ), (FrameChangeHandling, ),
+                    (FrameChangeHandling, ), (FrameChangeHandling, ), (FrameChangeHandling, ))
+}
