+++ feagi_data_structures/src/genomic/cortical_area/io_cortical_area_data_type_descriptors.rs
@@ -1,0 +1,13 @@
+
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum PercentageNeuronPositioning {
+    Linear,
+    Fractional,
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum FrameChangeHandling {
+    Absolute,
+    Incremental,
+}
