+++ feagi_data_structures/src/genomic/cortical_area/mod.rs
@@ -1,2 +1,6 @@
 mod cortical_area;
-mod cortical_area_types;
+mod input_output;
+mod io_cortical_area_data_type;
+mod io_cortical_area_data_type_descriptors;
+mod cortical_unit_type;
+mod cortical_type;
